# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "Diary API"
  description: MyDiary is an online journal where users can pen down their thoughts and feelings.
host: simondb.herokuapp.com
basePath: /
schemes:
- https
consumes:
- application/json
produces:
- application/json
paths:
  /api/v1/auth/signup:
    x-summary: User Signup
    post:
      summary: Allows a user to create an account
      responses:
        201:
          description: User Created
          examples:
            application/json:
              username: "peter"
              name: "Simon Peter"
              email: "kimbugwe@gmail.com"
              password: "12345678"
        400:
          description: Bad request
          examples:
            application/json:
              username: "peter"
              name: "Simon Peter"
              email: "kimbugwe@gmail.com"
  /api/v1/auth/login:
    x-summary: User Sign In
    post:
      summary: Allows a user to login into their account and provide a token for use
      responses:
        200:
          description: "logged in"
          examples:
            application/json:
              username: "peter"
              password: "12345678"
        401:
          description: Unauthorised
          examples:
            application/json:
              username: "peter"
              password: "1234568"
        400:
          description: Bad request
          examples:
            application/json:
              Message : "Parameter missing"
...